+++ cli/templates/starterApp/eslint.config.js
@@ -35,10 +35,10 @@ export default [
           },
         },
       ],
-      'comma-dangle': ['error', 'always-multiline'],
       'handle-callback-err': 'off',
       'consistent-return': 'off',
       'class-methods-use-this': 'off',
+      'comma-dangle': ['error', 'always-multiline'],
       'quotes': ['error', 'single', { allowTemplateLiterals: true }],
       'quote-props': ['error', 'consistent-as-needed'],
       'max-params': ['error', 4],
@@ -67,6 +67,7 @@ export default [
     },
     rules: {
       ...ts.configs.recommended.rules,
+      'no-undef': 'off',
     },
   },
 ];