+++ examples/todo-app/app/features/todos/list/todo/todo.model.js
@@ -1,3 +1,4 @@
+/** @implements {ITodoModel} */
 export class TodoModel {
   state = {
     title: '',