+++ CHANGELOG.md
@@ -2,11 +2,21 @@
 
 ## [Unreleased][unreleased]
 
+## [1.0.0-alpha.2][] - 2023-02-12
+
+### Changed
+
+- Improve starter application
+- Write learning comments
+- Change Telegram links
+- Update types
+
 ## [1.0.0-alpha.1][] - 2023-02-11
 
 ### Added
 
 - Implement initial engine features
 
-[unreleased]: https://github.com/rohiievych/swayer/compare/v1.0.0-alpha.1...HEAD
+[unreleased]: https://github.com/rohiievych/swayer/compare/v1.0.0-alpha.2...HEAD
+[1.0.0-alpha.2]: https://github.com/rohiievych/swayer/compare/v1.0.0-alpha.1...v1.0.0-alpha.2
 [1.0.0-alpha.1]: https://github.com/rohiievych/swayer/releases/tag/v1.0.0-alpha.1